Refactor storyChapterSelect and gameManager into scenes/
